requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.8
django-ninja==0.21.0
djongo
dnspython==2.3.0
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pydantic==1.8.2
pymongo==4.3.3
pytz==2023.3
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
